Fix incompatible pointer type warning on Scientific Linux Original: f0e0efaf0f98e2a52c6916567f760fe5c60f24ba

+++ lalinference/src/bayestar_distance.c
@@ -331,7 +331,7 @@ double bayestar_volume_render(
         double vec[3];
         cblas_dgemv(
             CblasRowMajor, CblasNoTrans, 3, 3, 1, R, 3, xyz, 1, 0, vec, 1);
-        long long ipix;
+        int64_t ipix;
         if (nest)
             vec2pix_nest64(nside, vec, &ipix);
         else
